[enhance]: token_id type int64->int32

+++ ginfer/model/qwen2_loader.cc
@@ -39,7 +39,7 @@ Qwen2Config Qwen2ModelLoader::loadConfig() {
   config.max_seq_len = 4096;  // TODO temporary fix for qwen2 models with longer context
   config.rms_norm_eps = json.value("rms_norm_eps", 1e-6f);
   config.rope_theta = json.value("rope_theta", 10000.0f);
-  config.eos_token_id = json.value("eos_token_id", static_cast<int64_t>(151645));
+  config.eos_token_id = json.value("eos_token_id", static_cast<int32_t>(151645));
 
   return config;
 }
